+++ SandboxiePlus/SandMan/Views/SbieView.cpp
@@ -610,6 +610,8 @@ void CSbieView::OnGroupAction()
 
 		// todo: fix behavioure on multiple selelction
 		foreach(const QString& Name, GetSelectedGroups(true)) {
+			bool bFound = false;
+			retry:
 			for (auto I = m_Groups.begin(); I != m_Groups.end(); ++I) {
 				int pos = I->indexOf(Name);
 				if (pos != -1) {
@@ -618,8 +620,14 @@ void CSbieView::OnGroupAction()
 						I.value()[pos+Offset] = I.value()[pos];
 						I.value()[pos] = Temp;
 					}
+					bFound = true;
 					break;
 				}
+			}
+			if (!bFound) {
+				bFound = true;
+				m_Groups[""].prepend(Name);
+				goto retry;
 			}
 		}
 	}
@@ -766,6 +774,11 @@ void CSbieView::OnSandBoxAction(QAction* Action)
 	}
 	else if (Action == m_pMenuBrowse)
 	{
+		if (SandBoxes.first()->IsEmpty()) {
+			QMessageBox("Sandboxie-Plus", tr("This Sandbox is empty."), QMessageBox::Information, QMessageBox::Ok, QMessageBox::NoButton, QMessageBox::NoButton, this).exec();
+			return;
+		}
+
 		CSandBoxPtr pBox = SandBoxes.first();
 		
 		static QMap<void*, CFileBrowserWindow*> FileBrowserWindows;